PR review changes included
