+++ components/VFBMain.js
@@ -12,11 +12,9 @@ import LinkButton from '../../../js/components/interface/linkButton/LinkButton';
 import ControlPanel from '../../../js/components/interface/controlPanel/controlpanel';
 import HTMLViewer from '../../../js/components/interface/htmlViewer/HTMLViewer';
 
-
 // import NewWidget from '../../../js/components/widgets/NewWidget'; For future use.
-require('./VFBMain.less');
 require('../css/base.less');
-require('../css/VFB.css');
+require('../css/VFBMain.less');
 //require('../css/base.css');
 var $ = require('jquery');
 var GEPPETTO = require('geppetto');
@@ -31,13 +29,13 @@ export default class VFBMain extends React.Component {
         this.state = {
             canvasAvailable: false,
             modelLoaded: (window.Model != undefined),
-            infoBtn: true, // All states handled by the button bar
-            stackBtn: true,
-            tutorialBtn: false,
-            searchBtn: true,
-            controlPanelBtn: true,
-            meshBtn: false,
-            queryBtn: true, // END buttonbar states
+            termInfoVisible: true, // All states handled by the button bar
+            stackViewerVisible: true,
+            tutorialWidgetVisible: false,
+            spotlightVisible: true,
+            controlPanelVisible: true,
+            wireframeVisible: false,
+            queryBuilderVisible: true, // END buttonbar states
             idSelected: undefined,
             htmlFromToolbar: undefined
         };
@@ -67,203 +65,11 @@ export default class VFBMain extends React.Component {
         this.htmlToolbarRef = this.htmlToolbarRef.bind(this);
 
         this.coli = 1;
-        this.colours = ["0x5b5b5b", "0x00ff00", "0xff0000", "0x0000ff", "0x0084f6", "0x008d46", "0xa7613e", "0x4f006a", "0x00fff6", "0x3e7b8d", "0xeda7ff", "0xd3ff95", "0xb94fff", "0xe51a58", "0x848400", "0x00ff95", "0x61002c", "0xf68412", "0xcaff00", "0x2c3e00", "0x0035c1", "0xffca84", "0x002c61", "0x9e728d", "0x4fb912", "0x9ec1ff", "0x959e7b", "0xff7bb0", "0x9e0900", "0xffb9b9", "0x8461ca", "0x9e0072", "0x84dca7", "0xff00f6", "0x00d3ff", "0xff7258", "0x583e35", "0x003e35", "0xdc61dc", "0x6172b0", "0xb9ca2c", "0x12b0a7", "0x611200", "0x2c002c", "0x5800ca", "0x95c1ca", "0xd39e23", "0x84b058", "0xe5edb9", "0xf6d3ff", "0xb94f61", "0x8d09a7", "0x6a4f00", "0x003e9e", "0x7b3e7b", "0x3e7b61", "0xa7ff61", "0x0095d3", "0x3e7200", "0xb05800", "0xdc007b", "0x9e9eff", "0x4f4661", "0xa7fff6", "0xe5002c", "0x72dc72", "0xffed7b", "0xb08d46", "0x6172ff", "0xdc4600", "0x000072", "0x090046", "0x35ed4f", "0x2c0000", "0xa700ff", "0x00f6c1", "0x9e002c", "0x003eff", "0xf69e7b", "0x6a7235", "0xffff46", "0xc1b0b0", "0x727272", "0xc16aa7", "0x005823", "0xff848d", "0xb08472", "0x004661", "0x8dff12", "0xb08dca", "0x724ff6", "0x729e00", "0xd309c1", "0x9e004f", "0xc17bff", "0x8d95b9", "0xf6a7d3", "0x232309", "0xff6aca", "0x008d12", "0xffa758", "0xe5c19e", "0x00122c", "0xc1b958", "0x00c17b", "0x462c00", "0x7b3e58", "0x9e46a7", "0x4f583e", "0x6a35b9", "0x72b095", "0xffb000", "0x4f3584", "0xb94635", "0x61a7ff", "0xd38495", "0x7b613e", "0x6a004f", "0xed58ff", "0x95d300", "0x35a7c1", "0x00009e", "0x7b3535", "0xdcff6a", "0x95d34f", "0x84ffb0", "0x843500", "0x4fdce5", "0x462335", "0x002c09", "0xb9dcc1", "0x588d4f", "0x9e7200", "0xca4684", "0x00c146", "0xca09ed", "0xcadcff", "0x0058a7", "0x2ca77b", "0x8ddcff", "0x232c35", "0xc1ffb9", "0x006a9e", "0x0058ff", "0xf65884", "0xdc7b46", "0xca35a7", "0xa7ca8d", "0x4fdcc1", "0x6172d3", "0x6a23ff", "0x8d09ca", "0xdcc12c", "0xc1b97b", "0x3e2358", "0x7b6195", "0xb97bdc", "0xffdcd3", "0xed5861", "0xcab9ff", "0x3e5858", "0x729595", "0x7bff7b", "0x95356a", "0xca9eb9", "0x723e1a", "0x95098d", "0xf68ddc", "0x61b03e", "0xffca61", "0xd37b72", "0xffed9e", "0xcaf6ff", "0x58c1ff", "0x8d61ed", "0x61b972", "0x8d6161", "0x46467b", "0x0058d3", "0x58dc09", "0x001a72", "0xd33e2c", "0x959546", "0xca7b00", "0x4f6a8d", "0x9584ff", "0x46238d", "0x008484", "0xf67235", "0x9edc84", "0xcadc6a", "0xb04fdc", "0x4f0912", "0xff1a7b", "0x7bb0d3", "0x1a001a", "0x8d35f6", "0x5800a7", "0xed8dff", "0x969696", "0xffd300"];
         this.vfbLoadBuffer = [];
         this.tutorialRender = undefined;
         this.htmlToolbarRender = undefined;
-        this.controlpanelRender = undefined;
-        this.querybuilderRender = undefined;
-        //this.logoStyle = { fontSize: '33px'};
-        this.tutorialsList = [
-            "/org.geppetto.frontend/geppetto/extensions/geppetto-vfb/tutorials/queryTutorial.json",
-            "/org.geppetto.frontend/geppetto/extensions/geppetto-vfb/tutorials/spotlightTutorial.json",
-            "/org.geppetto.frontend/geppetto/extensions/geppetto-vfb/tutorials/stackTutorial.json",
-            "/org.geppetto.frontend/geppetto/extensions/geppetto-vfb/tutorials/termTutorial.json"];
-
-        this.buttonBarConfig = {
-            "searchBtn": {
-                "icon": "fa fa-search",
-                "label": "",
-                "tooltip": "Search"
-            },
-            "queryBtn": {
-                "icon": "fa fa-quora",
-                "label": "",
-                "tooltip": "Open Query"
-            },
-            "controlPanelBtn": {
-                "icon": "fa fa-list",
-                "label": "",
-                "tooltip": "Layers"
-            },
-            "infoBtn": {
-                "icon": "fa fa-info",
-                "label": "",
-                "tooltip": "Show term info"
-            },
-            "stackBtn": {
-                "icon": "gpt-stack",
-                "label": "",
-                "tooltip": "Show slice viewer"
-            },
-            "meshBtn": {
-                "condition": "false",
-                "false": {
-                    "icon": "gpt-sphere_solid",
-                    "label": "",
-                    "tooltip": "Show wireframe"
-                },
-                "true": {
-                    "icon": "gpt-sphere_wireframe-jpg",
-                    "label": "",
-                    "tooltip": "Hide wireframe"
-                }
-            },
-            "tutorialBtn": {
-                "icon": "fa fa-question",
-                "label": "",
-                "tooltip": "Open tutorial"
-            }
-        };
-
-        this.spotlightConfig = {
-            "SpotlightBar": {
-                "DataSources": {},
-                "CompositeType": {
-                    "type": {
-                        "actions": [
-                            "window.setTermInfo($variableid$['$variableid$' + '_meta'],'$variableid$');GEPPETTO.Spotlight.close();",
-                        ],
-                        "icon": "fa-info-circle",
-                        "label": "Show info",
-                        "tooltip": "Show info"
-                    },
-                    "query": {
-                        actions: [
-                            "window.fetchVariableThenRun('$variableid$', window.addToQueryCallback);"
-                        ],
-                        icon: "fa-quora",
-                        label: "Add to query",
-                        tooltip: "Add to query"
-                    },
-                },
-                "VisualCapability": {
-                    "buttonOne": {
-                        "condition": "GEPPETTO.SceneController.isSelected($instances$)",
-                        "false": {
-                            "actions": ["GEPPETTO.SceneController.select($instances$)"],
-                            "icon": "fa-hand-stop-o",
-                            "label": "Unselected",
-                            "tooltip": "Select"
-                        },
-                        "true": {
-                            "actions": ["GEPPETTO.SceneController.deselect($instances$)"],
-                            "icon": "fa-hand-rock-o",
-                            "label": "Selected",
-                            "tooltip": "Deselect"
-                        },
-                    },
-                    "buttonTwo": {
-                        "condition": "GEPPETTO.SceneController.isVisible($instances$)",
-                        "false": {
-                            "actions": [
-                                "GEPPETTO.SceneController.show($instances$)"
-                            ],
-                            "icon": "fa-eye-slash",
-                            "label": "Hidden",
-                            "tooltip": "Show"
-                        },
-                        "true": {
-                            "actions": [
-                                "GEPPETTO.SceneController.hide($instances$)"
-                            ],
-                            "icon": "fa-eye",
-                            "label": "Visible",
-                            "tooltip": "Hide"
-                        }
-
-                    },
-                    "buttonThree": {
-                        "actions": [
-                            "GEPPETTO.SceneController.zoomTo($instances$);GEPPETTO.Spotlight.close();"
-                        ],
-                        "icon": "fa-search-plus",
-                        "label": "Zoom",
-                        "tooltip": "Zoom"
-                    },
-                }
-            }
-        };
-
-        this.spotlightDataSourceConfig = {
-            VFB: {
-                url: "http://solr.virtualflybrain.org/solr/ontology/select?fl=short_form,label,synonym,id,type,has_narrow_synonym_annotation,has_broad_synonym_annotation&start=0&fq=ontology_name:(vfb)&rows=250&bq=is_obsolete:false%5E100.0%20shortform_autosuggest:VFB*%5E110.0%20shortform_autosuggest:FBbt*%5E100.0%20is_defining_ontology:true%5E100.0%20label_s:%22%22%5E2%20synonym_s:%22%22%20in_subset_annotation:BRAINNAME%5E3%20short_form:FBbt_00003982%5E2&q=*$SEARCH_TERM$*%20OR%20$SEARCH_TERM$&defType=edismax&qf=label%20synonym%20label_autosuggest_ws%20label_autosuggest_e%20label_autosuggest%20synonym_autosuggest_ws%20synonym_autosuggest_e%20synonym_autosuggest%20shortform_autosuggest%20has_narrow_synonym_annotation%20has_broad_synonym_annotation&wt=json&indent=true", crossDomain: true,
-                crossDomain: true,
-                id: "short_form",
-                label: { field: "label", formatting: "$VALUE$" },
-                explode_fields: [{ field: "short_form", formatting: "$VALUE$ ($LABEL$)" }],
-                explode_arrays: [{ field: "synonym", formatting: "$VALUE$ ($LABEL$)" }],
-                type: {
-                    property: {
-                        icon: "fa-file-text-o",
-                        buttons: {
-                            buttonOne: {
-                                actions: ["window.addVfbId('$ID$');"],
-                                icon: "fa-info-circle",
-                                label: "Show info",
-                                tooltip: "Show info"
-                            }
-                        }
-                    },
-                    class: {
-                        icon: "fa-file-text-o",
-                        buttons: {
-                            buttonOne: {
-                                actions: ["window.addVfbId('$ID$');"],
-                                icon: "fa-info-circle",
-                                label: "Show info",
-                                tooltip: "Show info"
-                            },
-                            buttonTwo: {
-                                actions: ["window.fetchVariableThenRun('$ID$', window.addToQueryCallback, '$LABEL$');"],
-                                icon: "fa-quora",
-                                label: "Add to query",
-                                tooltip: "Add to query"
-                            }
-                        }
-                    },
-                    individual: {
-                        icon: "fa-file-image-o",
-                        buttons: {
-                            buttonOne: {
-                                actions: ["window.addVfbId('$ID$');"],
-                                icon: "fa-file-image-o",
-                                label: "Add to scene",
-                                tooltip: "Add to scene"
-                            },
-                            buttonTwo: {
-                                actions: ["window.fetchVariableThenRun('$ID$', window.addToQueryCallback, '$LABEL$');"],
-                                icon: "fa-quora",
-                                label: "Add to query",
-                                tooltip: "Add to query"
-                            }
-                        }
-                    }
-                },
-                bloodhoundConfig: {
-                    datumTokenizer: function (d) {
-                        return Bloodhound.tokenizers.nonword(d.label.replace('_', ' '));
-                    }.bind(this),
-                    queryTokenizer: function (q) {
-                        return Bloodhound.tokenizers.nonword(q.replace('_', ' '));
-                    }.bind(this),
-                    sorter: function (a, b) {
-                        var term = $('#typeahead').val();
-                        return this.customSorter(a, b, term);
-                    }.bind(this)
-                }
-            }
-        };
+        this.urlIdsLoaded = false;
+        this.idInitialized = false;
 
         this.stackConfiguration = {
             serverUrl: 'https://www.virtualflybrain.org/fcgi/wlziipsrv.fcgi',
@@ -280,320 +86,23 @@ export default class VFBMain extends React.Component {
             id: "StackViewerWidget_"
         };
 
-        this.controlPanelColMeta = [
-            {
-                "columnName": "path",
-                "order": 1,
-                "locked": false,
-                "displayName": "Path",
-                "source": "$entity$.getPath()"
-            },
-            {
-                "columnName": "name",
-                "order": 2,
-                "locked": false,
-                "customComponent": GEPPETTO.LinkComponent,
-                "displayName": "Name",
-                "source": "$entity$.getName()",
-                "actions": "window.addVfbId('$entity$');"
-            },
-            {
-                "columnName": "type",
-                "order": 3,
-                "locked": false,
-                "customComponent": GEPPETTO.LinkArrayComponent,
-                "displayName": "Type",
-                "source": "$entity$.$entity$_meta.getTypes().map(function (t) {return t.type.getInitialValue().value})",
-                "actions": "window.addVfbId('$entity$');",
-            },
-            {
-                "columnName": "controls",
-                "order": 4,
-                "locked": false,
-                "customComponent": GEPPETTO.ControlsComponent,
-                "displayName": "Controls",
-                "cssClassName": "controlpanel-controls-column",
-                "source": "",
-                "actions": "GEPPETTO.ControlPanel.refresh();"
-            },
-            {
-                "columnName": "image",
-                "order": 5,
-                "locked": false,
-                "customComponent": GEPPETTO.ImageComponent,
-                "displayName": "Image",
-                "cssClassName": "img-column",
-                "source": "GEPPETTO.ModelFactory.getAllVariablesOfMetaType($entity$.$entity$_meta.getType(), 'ImageType')[0].getInitialValues()[0].value.data"
-            }
-        ];
+        this.colours = require('./configuration/colours.json');
+        this.tutorialsList = require('./configuration/tutorialsList.json');
 
-        this.controlPanelConfig = {
-            "VisualCapability": {
-                "select": {
-                    "id": "select",
-                    "condition": "GEPPETTO.SceneController.isSelected($instance$.$instance$_obj != undefined ? [$instance$.$instance$_obj] : []) ||  GEPPETTO.SceneController.isSelected($instance$.$instance$_swc != undefined ? [$instance$.$instance$_swc] : [])",
-                    "false": {
-                        "actions": ["$instance$.select()"],
-                        "icon": "fa-hand-stop-o",
-                        "label": "Unselected",
-                        "tooltip": "Select",
-                        "id": "select",
-                    },
-                    "true": {
-                        "actions": ["$instance$.deselect()"],
-                        "icon": "fa-hand-rock-o",
-                        "label": "Selected",
-                        "tooltip": "Deselect",
-                        "id": "deselect",
-                    }
-                },
-                "color": {
-                    "id": "color",
-                    "actions": ["$instance$.setColor('$param$');"],
-                    "icon": "fa-tint",
-                    "label": "Color",
-                    "tooltip": "Color"
-                },
-                "zoom": {
-                    "id": "zoom",
-                    "actions": ["GEPPETTO.SceneController.zoomTo($instances$)"],
-                    "icon": "fa-search-plus",
-                    "label": "Zoom",
-                    "tooltip": "Zoom"
-                },
-                "visibility_obj": {
-                    "showCondition": "$instance$.getType().hasVariable($instance$.getId() + '_obj')",
-                    "condition": "(function() { var visible = false; if ($instance$.getType().$instance$_obj != undefined && $instance$.getType().$instance$_obj.getType().getMetaType() != GEPPETTO.Resources.IMPORT_TYPE && $instance$.$instance$_obj != undefined) { visible = GEPPETTO.SceneController.isVisible([$instance$.$instance$_obj]); } return visible; })()",
-                    "false": {
-                        "id": "visibility_obj",
-                        "actions": ["(function(){var instance = Instances.getInstance('$instance$.$instance$_obj'); if (instance.getType().getMetaType() == GEPPETTO.Resources.IMPORT_TYPE) { var col = instance.getParent().getColor(); instance.getType().resolve(function() { instance.setColor(col); GEPPETTO.trigger('experiment:visibility_changed', instance); GEPPETTO.ControlPanel.refresh(); }); } else { GEPPETTO.SceneController.show([instance]); }})()"],
-                        "icon": "gpt-shapehide",
-                        "label": "Hidden",
-                        "tooltip": "Show 3D Volume"
-                    },
-                    "true": {
-                        "id": "visibility_obj",
-                        "actions": ["GEPPETTO.SceneController.hide([$instance$.$instance$_obj])"],
-                        "icon": "gpt-shapeshow",
-                        "label": "Visible",
-                        "tooltip": "Hide 3D Volume"
-                    }
-                },
-                "visibility_swc": {
-                    "showCondition": "$instance$.getType().hasVariable($instance$.getId() + '_swc')",
-                    "condition": "(function() { var visible = false; if ($instance$.getType().$instance$_swc != undefined && $instance$.getType().$instance$_swc.getType().getMetaType() != GEPPETTO.Resources.IMPORT_TYPE && $instance$.$instance$_swc != undefined) { visible = GEPPETTO.SceneController.isVisible([$instance$.$instance$_swc]); } return visible; })()",
-                    "false": {
-                        "id": "visibility_swc",
-                        "actions": ["(function(){var instance = Instances.getInstance('$instance$.$instance$_swc'); if (instance.getType().getMetaType() == GEPPETTO.Resources.IMPORT_TYPE) { var col = instance.getParent().getColor(); instance.getType().resolve(function() { instance.setColor(col); GEPPETTO.trigger('experiment:visibility_changed', instance); GEPPETTO.ControlPanel.refresh(); }); } else { GEPPETTO.SceneController.show([instance]); }})()"],
-                        "icon": "gpt-3dhide",
-                        "label": "Hidden",
-                        "tooltip": "Show 3D Skeleton"
-                    },
-                    "true": {
-                        "id": "visibility_swc",
-                        "actions": ["GEPPETTO.SceneController.hide([$instance$.$instance$_swc])"],
-                        "icon": "gpt-3dshow",
-                        "label": "Visible",
-                        "tooltip": "Hide 3D Skeleton"
-                    }
-                },
-            },
-            "Common": {
-                "info": {
-                    "id": "info",
-                    "actions": ["var displayTxt = '$instance$'.split('.')['$instance$'.split('.').length - 1]; setTermInfo($instance$[displayTxt + '_meta'], displayTxt);"],
-                    "icon": "fa-info-circle",
-                    "label": "Info",
-                    "tooltip": "Info"
-                },
-                "delete": {
-                    "showCondition": "$instance$.getId()!=window.templateID",
-                    "id": "delete",
-                    "actions": ["if($instance$.parent != null){$instance$.parent.deselect();$instance$.parent.delete();}else{$instance$.deselect();$instance$.delete();};setTermInfo(window[window.templateID][window.templateID+'_meta'], window[window.templateID][window.templateID+'_meta'].getParent().getId());"],
-                    "icon": "fa-trash-o",
-                    "label": "Delete",
-                    "tooltip": "Delete"
-                }
-            }
-        };
+        this.buttonBarConfig = require('../components/configuration/buttonBarConfiguration').buttonBarConfig;
 
-        this.controlPanelControlConfigs = {
-            "Common": ['info', 'delete'],
-            "VisualCapability": ['select', 'color', 'visibility', 'zoom', 'visibility_obj', 'visibility_swc']
-        };
+        this.spotlightConfig = require('./configuration/spotlightConfiguration').spotlightConfig;
+        this.spotlightDataSourceConfig = require('./configuration/spotlightConfiguration').spotlightDataSourceConfig;
 
-        this.controlPanelColumns = ['name', 'type', 'controls', 'image'];
+        this.controlPanelColMeta = require('./configuration/controlPanelConfiguration').controlPanelColMeta;
+        this.controlPanelConfig = require('./configuration/controlPanelConfiguration').controlPanelConfig;
+        this.controlPanelControlConfigs = require('./configuration/controlPanelConfiguration').controlPanelControlConfigs;
+        this.controlPanelColumns = require('./configuration/controlPanelConfiguration').controlPanelColumns;
 
-        this.queryResultsColMeta = [
-            {
-                "columnName": "id",
-                "order": 1,
-                "locked": false,
-                "visible": true,
-                "displayName": "ID",
-            },
-            {
-                "columnName": "name",
-                "order": 2,
-                "locked": false,
-                "visible": true,
-                "customComponent": GEPPETTO.QueryLinkComponent,
-                "actions": "window.addVfbId('$entity$');",
-                "displayName": "Name",
-                "cssClassName": "query-results-name-column",
-            },
-            {
-                "columnName": "description",
-                "order": 3,
-                "locked": false,
-                "visible": true,
-                "displayName": "Definition",
-                "cssClassName": "query-results-description-column"
-            },
-            {
-                "columnName": "type",
-                "order": 4,
-                "locked": false,
-                "visible": true,
-                "displayName": "Type",
-                "cssClassName": "query-results-type-column"
-            },
-            {
-                "columnName": "controls",
-                "order": 5,
-                "locked": false,
-                "visible": false,
-                "customComponent": GEPPETTO.QueryResultsControlsComponent,
-                "displayName": "Controls",
-                "actions": "",
-                "cssClassName": "query-results-controls-column"
-            },
-            {
-                "columnName": "images",
-                "order": 6,
-                "locked": false,
-                "visible": true,
-                "customComponent": GEPPETTO.SlideshowImageComponent,
-                "displayName": "Images",
-                "actions": "window.addVfbId('$entity$');",
-                "cssClassName": "query-results-images-column"
-            },
-            {
-                "columnName": "score",
-                "order": 7,
-                "locked": false,
-                "visible": true,
-                "displayName": "Score",
-                "cssClassName": "query-results-score-column"
-            }
-        ];
-
-        // which columns to display in the results
-        this. queryResultsColumns = ['name', 'description', 'type', 'images', 'score'];
-
-        this.queryResultsControlConfig = {
-            "Common": {
-                "info": {
-                    "id": "info",
-                    "actions": [
-                        "window.addVfbId('$ID$');"
-                    ],
-                    "icon": "fa-info-circle",
-                    "label": "Info",
-                    "tooltip": "Info"
-                },
-                "flybase": {
-                    "showCondition": "'$ID$'.startsWith('FBbt')",
-                    "id": "flybase",
-                    "actions": [
-                        "window.open('http://flybase.org/cgi-bin/cvreport.html?rel=is_a&id=' + '$ID$'.replace(/_/g, ':'), '_blank').focus()"
-                    ],
-                    "icon": "gpt-fly",
-                    "label": "FlyBase",
-                    "tooltip": "FlyBase Term"
-                },
-                "flybase": {
-                    "showCondition": "('$ID$'.startsWith('FB') && !'$ID$'.startsWith('FBbt'))",
-                    "id": "flybase",
-                    "actions": [
-                        "window.open('http://flybase.org/reports/' + '$ID$'.replace(/_/g, ':'), '_blank').focus()"
-                    ],
-                    "icon": "gpt-fly",
-                    "label": "FlyBase",
-                    "tooltip": "FlyBase Report"
-                }
-            }
-        };
-
-        // add datasource config to query control
-        this.queryBuilderDatasourceConfig = {
-            VFB: {
-                url: "https://solr.virtualflybrain.org/solr/ontology/select?fl=short_form,label,synonym,id,type,has_narrow_synonym_annotation,has_broad_synonym_annotation&start=0&fq=ontology_name:(vfb)&rows=250&bq=is_obsolete:false%5E100.0%20shortform_autosuggest:VFB*%5E100.0%20shortform_autosuggest:FBbt*%5E100.0%20is_defining_ontology:true%5E100.0%20label_s:%22%22%5E2%20synonym_s:%22%22%20in_subset_annotation:BRAINNAME%5E3%20short_form:FBbt_00003982%5E2&q=*$SEARCH_TERM$*%20OR%20$SEARCH_TERM$&defType=edismax&qf=label%20synonym%20label_autosuggest_ws%20label_autosuggest_e%20label_autosuggest%20synonym_autosuggest_ws%20synonym_autosuggest_e%20synonym_autosuggest%20shortform_autosuggest%20has_narrow_synonym_annotation%20has_broad_synonym_annotation&wt=json&indent=true", crossDomain: true,
-                crossDomain: true,
-                id: "short_form",
-                label: { field: "label", formatting: "$VALUE$" },
-                explode_fields: [{ field: "short_form", formatting: "$VALUE$ ($LABEL$)" }],
-                explode_arrays: [{ field: "synonym", formatting: "$VALUE$ ($LABEL$)" }],
-                type: {
-                    class: {
-                        actions: ["window.fetchVariableThenRun('$ID$', function(){ GEPPETTO.QueryBuilder.addQueryItem({ term: '$LABEL$', id: '$ID$'}); });"],
-                        icon: "fa-dot-circle-o"
-                    },
-                    individual: {
-                        actions: ["window.fetchVariableThenRun('$ID$', function(){ GEPPETTO.QueryBuilder.addQueryItem({ term: '$LABEL$', id: '$ID$'}); });"],
-                        icon: "fa-square-o"
-                    }
-                },
-                queryNameToken: '$NAME',
-                resultsFilters: {
-                    getItem: function (record, header, field) {
-                        var recordIndex = header.indexOf(field);
-                        return record[recordIndex]
-                    },
-                    getId: function (record) {
-                        return record[0]
-                    },
-                    getName: function (record) {
-                        return record[1]
-                    },
-                    getDescription: function (record) {
-                        return record[2]
-                    },
-                    getType: function (record) {
-                        return record[3]
-                    },
-                    getImageData: function (record) {
-                        return record[4]
-                    },
-                    getScore: function (record) {
-                        return record[5]
-                    },
-                    getRecords: function (payload) {
-                        return payload.results.map(function (item) {
-                            return item.values
-                        })
-                    },
-                    getHeaders: function (payload) {
-                        return payload.header;
-                    }
-                },
-                bloodhoundConfig: {
-                    datumTokenizer: function (d) {
-                        return Bloodhound.tokenizers.nonword(d.label.replace('_', ' '));
-                    }.bind(this),
-                    queryTokenizer: function (q) {
-                        return Bloodhound.tokenizers.nonword(q.replace('_', ' '));
-                    }.bind(this),
-                    sorter: function (a, b) {
-                        var term = $("#query-typeahead").val();
-                        return this.customSorter(a, b, term);
-                    }.bind(this)
-                }
-            }
-        };
-
-        this.urlIdsLoaded = false;
-        this.idInitialized = false;
+        this.queryResultsColMeta = require('./configuration/queryBuilderConfiguration').queryResultsColMeta;
+        this.queryResultsColumns = require('./configuration/queryBuilderConfiguration').queryResultsColumns;
+        this.queryResultsControlConfig = require('./configuration/queryBuilderConfiguration').queryResultsControlConfig;
+        this.queryBuilderDatasourceConfig = require('./configuration/queryBuilderConfiguration').queryBuilderDatasourceConfig;
 
         window.redirectURL = '$PROTOCOL$//$HOST$/?i=$TEMPLATE$,$VFB_ID$&id=$VFB_ID$';
     }
@@ -607,7 +116,7 @@ export default class VFBMain extends React.Component {
     };
 
     getStackViewerDefaultX() {
-        return (window.innerWidth - (Math.ceil(window.innerWidth / 4) + 10)); 
+        return (window.innerWidth - (Math.ceil(window.innerWidth / 4) + 10));
     };
 
     getStackViewerDefaultY() {
@@ -791,7 +300,7 @@ export default class VFBMain extends React.Component {
 
     fetchVariableThenRun(variableId, callback, label) {
         GEPPETTO.SceneController.deselectAll(); // signal something is happening!
-        var variables = GEPPETTO.ModelFactory.getTopLevelVariablesById([variableId]);
+        var variables = GEPPETTO.ModelFactory.getTopLevelVariablesById(variableId);
         if (!variables.length > 0) {
             Model.getDatasources()[0].fetchVariable(variableId, function () {
                 if (callback != undefined)
@@ -983,27 +492,27 @@ export default class VFBMain extends React.Component {
 
     componentDidUpdate(prevProps, prevState) {
         // Reopen Terminfo from button bar if previously has been closed.
-        if((this.state.infoBtn === true) && (prevState.infoBtn !== this.state.infoBtn) && (prevState.infoBtn !== undefined)) {
+        if((this.state.termInfoVisible === true) && (prevState.termInfoVisible !== this.state.termInfoVisible) && (prevState.termInfoVisible !== undefined)) {
             this.refs.termInfoRef.addTermInfo();
         }
         // Reopen stackViewer from button bar if previously has been closed.
-        if((prevState.stackBtn !== this.state.stackBtn) && (prevState.stackBtn !== undefined)) {
+        if((prevState.stackViewerVisible !== this.state.stackViewerVisible) && (prevState.stackViewerVisible !== undefined)) {
             this.refs.stackViewerRef.addStackWidget();
         }
 
-        if((prevState.meshBtn !== this.state.meshBtn)) {
+        if((prevState.wireframeVisible !== this.state.wireframeVisible)) {
             this.refs.vfbCanvas.setWireframe(!this.refs.vfbCanvas.getWireframe());
         }
 
-        if((prevState.controlPanelBtn !== this.state.controlPanelBtn)) {
+        if((prevState.controlPanelVisible !== this.state.controlPanelVisible)) {
             $('#controlpanel').show();
         }
 
-        if((prevState.searchBtn !== this.state.searchBtn)) {
+        if((prevState.spotlightVisible !== this.state.spotlightVisible)) {
             $('#spotlight').show();
         }
 
-        if((prevState.queryBtn !== this.state.queryBtn)) {
+        if((prevState.queryBuilderVisible !== this.state.queryBuilderVisible)) {
             $('#querybuilder').show();
         }
 
@@ -1190,7 +699,7 @@ export default class VFBMain extends React.Component {
 
     // Children handlers
     buttonBarHandler(buttonState) {
-        if((buttonState !== "infoBtn") || (buttonState !== "stackBtn")) {
+        if((buttonState !== "termInfoVisible") || (buttonState !== "stackViewerVisible")) {
             var tempState = this.state[buttonState];
             this.setState({
                 [buttonState] : !tempState
@@ -1199,26 +708,26 @@ export default class VFBMain extends React.Component {
     }
 
     tutorialHandler() {
-        if(this.state.tutorialBtn == true) {
+        if(this.state.tutorialWidgetVisible == true) {
             this.setState({
-                tutorialBtn: false
+                tutorialWidgetVisible: false
             });
         }
     }
 
     termInfoHandler(childrenState) {
         if(childrenState !== undefined) {
-            this.setState({ infoBtn: childrenState });
+            this.setState({ termInfoVisible: childrenState });
         } else {
-            this.setState({ infoBtn: !(this.state.infoBtn) });
+            this.setState({ termInfoVisible: !(this.state.termInfoVisible) });
         }       
     };
 
     stackViewerHandler(childrenState) {
         if(childrenState !== undefined) {
-            this.setState({ stackBtn: childrenState });
+            this.setState({ stackViewerVisible: childrenState });
         } else {
-            this.setState({ stackBtn: !(this.state.infoBtn) });
+            this.setState({ stackViewerVisible: !(this.state.stackViewerVisible) });
         }
     };
 
@@ -1250,9 +759,9 @@ export default class VFBMain extends React.Component {
 
     render() {
 
-        if((this.state.tutorialBtn == true) && (this.tutorialRender == undefined)) {
+        if((this.state.tutorialWidgetVisible == true) && (this.tutorialRender == undefined)) {
             this.tutorialRender = <TutorialWidget tutorialHandler={this.tutorialHandler} ref="tutorialWidgetRef" />
-        } else if((this.state.tutorialBtn == true) && (this.tutorialRender != undefined)){
+        } else if((this.state.tutorialWidgetVisible == true) && (this.tutorialRender != undefined)){
             this.refs.tutorialWidgetRef.refs.tutorialRef.open(true);
         }
