--- components/VFBMain.less
@@ -1,49 +1,0 @@
-@import "../css/colors.less";
-
-@-webkit-keyframes fadeIn {
-		from {
-			opacity: 0;
-		}
-		to {
-			opacity: 1;
-		}
-	}
-
-	@keyframes fadeIn {
-		from {
-			opacity: 0;
-		}
-		to {
-			opacity: 1;
-		}
-	}
-
-#geppettologo {
-	top:55px;
-}
-
-.htmlViewerVFB {
-	animation: fadeIn 1s;
-	-webkit-animation: fadeIn 1s;
-	transition: opacity 1s ease-out;
-	background-color: @background_color_widget;
-	background: @background_color_widget;
-	color: white;
-	opacity: 0.9;
-	z-index: 999999;
-	overflow-y: scroll;
-	box-shadow:0 0 0 1600px rgba(0,0,0,0.65);
-}
-
-.closeHTMLViewer {
-	float: right;
-	height: 5px;
-	font-size: 26px;
-	color: @primary_color;
-	margin-left: 5px;
-	cursor: pointer;
-}
-
-#HTMLViewerContainer {
-	padding: 20px;
-}
