+++ components/configuration/tutorialsList.json
@@ -1,0 +1,6 @@
+[
+    "/org.geppetto.frontend/geppetto/extensions/geppetto-vfb/tutorials/queryTutorial.json",
+    "/org.geppetto.frontend/geppetto/extensions/geppetto-vfb/tutorials/spotlightTutorial.json",
+    "/org.geppetto.frontend/geppetto/extensions/geppetto-vfb/tutorials/stackTutorial.json",
+    "/org.geppetto.frontend/geppetto/extensions/geppetto-vfb/tutorials/termTutorial.json"
+]
