--- css/VFB.css
@@ -1,60 +1,0 @@
-.spinner-container .gpt-fly {
-  width: 100px;
-  font-size: 100px;
-  margin: 0 auto;
-  margin-bottom: 30px;
-}
-
-.transparent {
-    background: transparent !important;
-}
-
-#querybuilder .query-results-images-column img{
-	max-height: 110px;
-}
-
-#querybuilder .query-results-images-column img:hover{
-	max-height: 300px;
-}
-
-.tutorialMessage{
-    height : 100% !important;
-}
-
-.gpt-query:before {
-    margin-top: 3px;
-}
-
-.gpt-shapehide:before {
-    margin-top: 3px;
-}
-
-.gpt-shapeshow:before {
-    margin-top: 5px;
-}
-
-.ui-widget {
-    font-family: arial, sans-serif; 
-    font-size:13px; 
-}
-
-@media screen and (max-width: 1200px) {
-    .uk-visible-large {
-        display: none !important;
-    }
-}
-
-@media screen and (max-width: 1000px) {
-    .wideDivC {
-        display: none !important;
-    }
-}
-
-@media screen and (max-width: 680px) {
-    .uk-navbar-content {
-        display: none !important;
-    }
-}
-
-
-
